backend/paddleocr/requirements.txt
paddlepaddle
paddleocr
PyMuPDF
Pillow
